Database enhancement (#2)
Databases with ActiveRecord are now supported in a `development.sqlite3` database file. There is no production or testing database configured, but when launching with Fly, you can `.dockerignore` the `db/development.sqlite3` file and also add `RUN bundle exec rake db:migrate` to the `Dockerfile` in order to create a fresh new database on the deployed app. Not best practice, but simple and works well.